+++ src/pystackapi/_base_client.py
@@ -8,7 +8,7 @@ from ._raw_response_dict import RawResponseDict
 
 class BaseClient:
     """Base class for clients of the StackExchange API."""
-    def _call(self, url: str, params: dict) -> RawResponseDict:
+    def _get(self, url: str, params: dict) -> RawResponseDict:
         req_params = ('?' if params else '') + '&'.join(f'{k}={v}' for k, v in params.items())
 
         response = requests.get(f'{url}{req_params}')
@@ -17,4 +17,12 @@ class BaseClient:
             raise HttpError(response.status_code, response.json())
 
         # we guarantee that `response.json` is `RawResponseDict`.
+        return cast(RawResponseDict, response.json())
+
+    def _post(self, url: str, data: dict) -> RawResponseDict:
+        response = requests.post(url, data=data)
+
+        if response.status_code != 200:
+            raise HttpError(response.status_code, response.json())
+
         return cast(RawResponseDict, response.json())