+++ tests/test_base_client/__init__.py
@@ -1,3 +1,4 @@
 """Tests for class `BaseClient`."""
 
-from . import test_call
+from . import test_get
+from . import test_post