+++ tests/test_base_client/test_get.py
@@ -1,4 +1,4 @@
-"""Tests for method `BaseClient._call`."""
+"""Tests for method `BaseClient._get`."""
 import lest
 
 from pystackapi import _base_client as client_m
@@ -19,22 +19,22 @@ def reset_requests() -> None:
 
 
 @lest.register
-def test_simple_call_url() -> None:
-    base_client._call('http://somesite.org/somepage.html', {})
+def test_simple_get_url() -> None:
+    base_client._get('http://somesite.org/somepage.html', {})
 
     lest.assert_eq(requests.url, 'http://somesite.org/somepage.html')
 
 
 @lest.register
-def test_call_with_one_arg_url() -> None:
-    base_client._call('http://somesite.org/somepage.html', {'arg1': 'hello'})
+def test_get_with_one_arg_url() -> None:
+    base_client._get('http://somesite.org/somepage.html', {'arg1': 'hello'})
 
     lest.assert_eq(requests.url, 'http://somesite.org/somepage.html?arg1=hello')
 
 
 @lest.register
-def test_call_with_many_args_url() -> None:
-    base_client._call('http://somesite.org/somepage.html', {'arg1': 'a', 'arg2': 'b', 'arg3': 'c'})
+def test_get_with_many_args_url() -> None:
+    base_client._get('http://somesite.org/somepage.html', {'arg1': 'a', 'arg2': 'b', 'arg3': 'c'})
 
     lest.assert_eq(requests.url, 'http://somesite.org/somepage.html?arg1=a&arg2=b&arg3=c')
 
@@ -47,13 +47,13 @@ def test_handling_error() -> None:
     l_base_client = BaseClient()
 
     with lest.assert_raises(HttpError):
-        l_base_client._call('http://somesite.org/somepage.html', {})
+        l_base_client._get('http://somesite.org/somepage.html', {})
 
     client_m.__dict__['requests'] = requests  # reset
 
 
 @lest.register
-def test_call_return_value() -> None:
-    res = base_client._call('http://somesite.org/somepage.html', {})
+def test_get_return_value() -> None:
+    res = base_client._get('http://somesite.org/somepage.html', {})
 
     lest.assert_eq(res, {'items': [{'id': 1}]})